PS2CONFIG
- Updated PS2 config for Secret Agent Clank submitted by mrjaredbeta
Fixes run/walk miscalculations by 0x0E patches and fixes the "robot falling through the floor" bug that breaks the game with 0x26 patch (there's 4 math opcodes but did the lazy Sony way of patching whole function instead). This may or may not be the final config since the save file I used didn't have the last level unlocked (I think), but everything else should be fixed. :)

The order of the 0x0E patches in the config are: Clank missions, Ratchet missions, Qwark missions, and Clank's special mission. And maybe the max amount of 0x0E is 32? I didn't attempt more, but this is not breaking netemu even with 32 0x0E commands.