PS2CONFIG
- Updated configs for Radiata Stories fixed by Agrippa
Fixed frame rate by disabling the shadows. NTSC-J version untested. Original codes found by Maori-Jigglypuff. Some small glitches are still present. First of all, similarly to the PS4, camera after the battle is not positioned on the characters. Second, I have noticed a small glitch with the Jack's clothes at the end of the first meeting in the Ganz room with Ridley. And the last but not least, there is a static noise popping at the beginning of the new game's intro (similar to the VP2 Lost Forest one, probably caused by the A/V sync and skipping samples).