PS2CONFIG
- Added ps2 configs for GT4 Prologue fixed by Agrippa
Enable pressure sensitive controls and fix the stuttering of videos. Jitter on the car selection screen, regardless if the shadows are disabled or not.

- Added ps2 configs for Champions: Return to Arms fixed by Agrippa
Disabled heat haze effect to improve the frame rate in the Lake of Fire maps. Still slowdowns do happen there, especially in the NTSC version. Zooming in the camera fully, near the fire shamans, seems to increase the slowdown in the NTSC version.