ps3netsrv 20220228
- Improved performance of vISO reading large files (PKG, PSARC, Videos, etc)
VisoFile::read() in the original code open/close the file every time the buffer is filled.
Now the file remains open, until a different file is read. This change significantly reduces the I/O overhead reading very large files.
- memcpy now uses 64bit buffer when size is multiple of 8.
It is expected that this increase the performance copying buffers in 8X.