PS2CONFIG
- Removed alternative PS2 config for WRC: World Rally Championship: Rally Evolved (remapped buttons)
- Updated PS2 configs for Gran Turismo 4, fixed by Agrippa
Disables rendering of light sources to fix the flickering of them and to improve the frame rate on night tracks.
- Added alternative PS2 config for Valkyrie Profile 2: Silmeria PAL Progressive and Hor+, fix by Agrippa
- enables the leftover full height front buffer mode from the NTSC versions of game,
- fixes the vert- widescreen in the 2D mode and real-time cut scenes,
- should disable the anti-tamper memory scanning completely (improved @Maori-Jigglypuff original patch by me),
- retains frame rate and anti-ghosting patches from the CUSTOM config.
The warp effect had to be disabled to prevent crashing with the hi-res buffer enabled. New anti-tamper patches are highly experimental. Battle transition screen is glitched one frame long. Prerecorded videos are out of sync (the same issue as 480p mode in the NTSC version).