PS2CONFIG
- Added PS2 config for Tokyo Bus Annai: Kyou kara Kimi mo Untenshu, fixed by mrjaredbeta
Ensures DMA transfer is completed to avoid graphical corruption and flickering.
- Updated PS2 config for Evergrace, fixed by mrjaredbeta
Ensures DMA transfer is completed to avoid graphical corruption and flickering.
- Updated PS2 config for Gitaroo Man, fixed by mrjaredbeta
The method from above causes too much of a performance decrease, so the original implementation is needed. There can still be occasionally flickering on some levels, but it goes largely unnoticed.