PS2CONFIG
- Black (SLUS-21376) (Proof of Concept Performance Fix) by mrjaredbeta
Here is a fix that limits frame rate to 20fps and speeds up animations a bit to retain full speed gameplay. Not the most "elegant" way to go about it, but game speed is at least 100% throughout now. The eventual fix would be to trigger this only when frame rate/game speed drops in order to have 30fps the majority of the time. Apparently, the game also has some other issues with teleportation (even with the current COP2 instruction patch), but have not been tested far enough to experience it.